+++ src/components/Navigation/Navigation.vue
@@ -1,73 +1,76 @@
 <template>
-  <nav :class="$style.nav">
-    <search :class="$style.search" />
-    <list :class="$style.list" />
-    <div :class="$style.footer">
-      <span :class="$style.version">booQ Project v{{ version }}</span>
-      <button :class="$style.aboutButton" @click="onClickAbout">?</button>
-    </div>
-  </nav>
+  <transition name="navigation">
+    <navigation-content
+      v-if="!canToggleNavigationShown"
+      :class="$style.desktopNavigation"
+    />
+  </transition>
+  <template v-if="canToggleNavigationShown">
+    <transition name="fade-fast">
+      <div
+        v-show="isNavigationShown"
+        :class="$style.mobileNavigationDim"
+        @click="emit('toggleNavigationShown')"
+      ></div>
+    </transition>
+    <transition name="navigation">
+      <navigation-content
+        v-show="isNavigationShown"
+        :class="$style.mobileNavigation"
+      />
+    </transition>
+  </template>
 </template>
 
 <script lang="ts">
 import { defineComponent } from 'vue'
-import { useRouter } from 'vue-router'
-import version from '/@/lib/version'
-import Search from './Search.vue'
-import List from './List.vue'
+import NavigationContent from './NavigationContent.vue'
 
 export default defineComponent({
   name: 'Navigation',
   components: {
-    Search,
-    List
+    NavigationContent
   },
-  setup() {
-    const router = useRouter()
-
-    const onClickAbout = () => {
-      router.push('/about')
+  props: {
+    canToggleNavigationShown: {
+      type: Boolean,
+      required: true
+    },
+    isNavigationShown: {
+      type: Boolean,
+      required: true
     }
-
-    return { onClickAbout, version }
+  },
+  emits: {
+    toggleNavigationShown: () => true
+  },
+  setup(props, { emit }) {
+    return { emit }
   }
 })
 </script>
 
 <style lang="scss" module>
-.nav {
-  display: flex;
-  flex-direction: column;
-  padding: 1rem;
-  background-color: $color-secondary;
-  > * {
-    margin: 1rem 0;
-    &:first-child {
-      margin-top: 0;
-    }
-    &:last-child {
-      margin-bottom: 0;
-    }
-  }
+.desktopNavigation {
+  flex-shrink: 0;
+  width: 260px;
 }
-.search {
-  width: auto;
+.mobileNavigation {
+  position: absolute;
+  top: 0;
+  left: 0;
+  right: 0;
+  bottom: 0;
+  width: 260px;
+  z-index: 1;
 }
-.list {
-  flex: 1;
-}
-.footer {
-  display: flex;
-  align-items: center;
-  text-align: left;
-}
-.version {
-  flex: 1;
-  margin-right: 1rem;
-}
-.aboutButton {
-  width: 2em;
-  height: 2em;
-  cursor: pointer;
+.mobileNavigationDim {
+  position: absolute;
+  top: 0;
+  left: 0;
+  right: 0;
+  bottom: 0;
+  z-index: 1;
+  background: $color-background-dim;
 }
 </style>
